Resolve launcher activity in `AndroidExt`
