+++ maestro-client/src/main/java/maestro/drivers/AndroidDriver.kt
@@ -35,6 +35,7 @@ import maestro.SwipeDirection
 import maestro.TreeNode
 import maestro.android.AndroidAppFiles
 import maestro.android.asManifest
+import maestro.android.resolveLauncherActivity
 import maestro_android.MaestroDriverGrpc
 import maestro_android.deviceInfoRequest
 import maestro_android.tapRequest
@@ -153,37 +154,17 @@ class AndroidDriver(
         try {
             val apkFile = AndroidAppFiles.getApkFile(dadb, appId)
             val manifest = apkFile.asManifest()
-            if (hasThirdPartyLauncherConfigured(manifest, appId) && hasOnlyAppLauncher(manifest, appId)) {
-                val activity = manifest.activities.first {
-                    it.isHomeActivity &&
-                        it.name.split(".").intersect(appId.split(".").toSet()).isNotEmpty()
-                }
-                shell("am start-activity -n $appId/${activity.name}")
-            } else {
-                try {
-                    shell("am start-activity -n $appId/${manifest.launcherActivity.name}")
-                } catch (exception: Exception) {
+            val launcherActivity = manifest.resolveLauncherActivity(appId)
+            runCatching { shell("am start-activity -n $appId/${launcherActivity}") }
+                .onFailure {
                     shell("monkey --pct-syskeys 0 -p $appId 1")
                 }
-            }
         } catch (ioException: IOException) {
             shell("monkey --pct-syskeys 0 -p $appId 1")
         } catch (saxException: SAXException) {
             shell("monkey --pct-syskeys 0 -p $appId 1")
         }
     }
-
-    private fun hasOnlyAppLauncher(manifest: ManifestData, appId: String) =
-        manifest.activities.filter {
-            it.isHomeActivity &&
-                it.name.split(".").intersect(appId.split(".").toSet()).isNotEmpty()
-        }.size == 1
-
-    private fun hasThirdPartyLauncherConfigured(manifest: ManifestData, appId: String) =
-        manifest.activities.any {
-            it.isHomeActivity &&
-                it.name.split(".").intersect(appId.split(".").toSet()).isEmpty()
-        }
 
     override fun stopApp(appId: String) {
         // Note: If the package does not exist, this call does *not* throw an exception
